Moved constants to root dir | Now defining adaptFetchUrl in index.js | Closes #8
