+++ index.js
@@ -1,3 +1,7 @@
-export { IS_LOCAL_WEB_APP } from "./modules/constants.js";
+import { IS_LOCAL_WEB_APP, WEB_APP_DEV_PROXY_URL } from "./constants";
 
-export { adaptFetchUrl } from "./modules/utils.js";
+export { IS_LOCAL_WEB_APP };
+
+export const adaptFetchUrl = (url) => {
+	return IS_LOCAL_WEB_APP ? `${WEB_APP_DEV_PROXY_URL}/${url}` : url;
+};
