--- modules/utils.js
@@ -1,5 +1,0 @@
-import { IS_LOCAL_WEB_APP, WEB_APP_DEV_PROXY_URL } from "./constants";
-
-export const adaptFetchUrl = (url) => {
-	return IS_LOCAL_WEB_APP ? `${WEB_APP_DEV_PROXY_URL}/${url}` : url;
-};
